Add s390x arch support

+++ hack/init-buildx.sh
@@ -32,7 +32,8 @@ fi
 # Ensure qemu is in binfmt_misc
 # Docker desktop already has these in versions recent enough to have buildx
 # We only need to do this setup on linux hosts
-if [ "$(uname)" == 'Linux' ]; then
+# We don't need qemu in s390x, since we won't cross compile
+if [ "$(uname)" == 'Linux' ] && [ ${ARCH} != "s390x" ]; then
   # NOTE: this is pinned to a digest for a reason!
   # Note2 (@rikatz) - Removing the pin, as apparently it's breaking new alpine builds
   # docker run --rm --privileged multiarch/qemu-user-static@sha256:28ebe2e48220ae8fd5d04bb2c847293b24d7fbfad84f0b970246e0a4efd48ad6 --reset -p yes
@@ -54,7 +55,8 @@ fi
 # Ensure qemu is in binfmt_misc
 # Docker desktop already has these in versions recent enough to have buildx
 # We only need to do this setup on linux hosts
-if [ "$(uname)" == 'Linux' ]; then
+# We don't need qemu in s390x, since we won't cross compile
+if [ "$(uname)" == 'Linux' ] && [ ${ARCH} != "s390x" ]; then
   # NOTE: this is pinned to a digest for a reason!
   # https://github.com/docker/buildx/issues/542#issuecomment-778835576
   docker run --rm --privileged tonistiigi/binfmt --uninstall qemu-aarch64 && docker run --rm --privileged tonistiigi/binfmt --install arm64
